requirements.txt
httpx>=0.27.0
websockets>=12.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
